fixed bug with card reads not working after using web button
