+++ unlock_door.py
@@ -1,4 +1,29 @@
-from read_process import unlock_door
+from __future__ import print_function
+
+from RPIO import PWM
+import time
+
+from read_process import log
+
+servo_pin = 27
+servo_range = [600, 1300]  # Left, Right (2300 Max, 500 Min)
+
+PWM.set_loglevel(PWM.LOG_LEVEL_ERRORS)
+servo = PWM.Servo()
+
+def unlock_door(method='card'):
+    if method == 'button':
+        log('True', 'Web Button', 'Web User')
+    print('Unlocking...')
+    servo.set_servo(servo_pin, servo_range[1])
+    time.sleep(4)
+    print('Locking...')
+    servo.set_servo(servo_pin, servo_range[0])
+    time.sleep(2)
+    servo.stop_servo(servo_pin)
 
 if __name__ == '__main__':
-    unlock_door(method='button')
+    try:
+        unlock_door(method='button')
+    except:
+        servo.stop_servo(servo_pin)
